backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
psycopg2==2.9.9
sqlparse==0.5.0
environs==11.0.0
gunicorn==22.0.0
